LQR model for positional and velocity control of the AUV
The .m file consists of the MATLAB code to initiate the Simulink models. The three Simulink models are of the LQR based controller design on the states of the AUV, where noise is introduced using a Gaussian noise block  in two subsequent versions; where the noise is reduced using a gain block in one and an LPF in another.